chore: 完成项目大规模重构与依赖更新
大量变更说明：
1.  清理冗余的 alembic 初始迁移文件与空目录
2.  移除废弃的 SVM 大奖判断相关代码与模型文件
3.  更新 alembic 脚本模板与多数据库配置，支持数据迁移分离
4.  重构第三方抽奖动态模块，迁移配置到 CONFIG 全局设置
5.  删除 t_lotdyninfo 中废弃的 isFollowed 字段
6.  新增 rawJsonStr 数据回填脚本与批量接口
7.  新增请求耗时统计中间件与爬虫卡死检测任务
8.  更新依赖包，添加 vllm、jsonlines 等新依赖
9.  优化数据库错误处理逻辑，添加死锁异常抛出
10. 新增批量提交抽奖动态接口与时间筛选扩展
11. 修复 grpc 脚本路径与日志输出问题
12. 重构代理模型字段类型，优化数据库索引

+++ FastapiApp/Utils/代理/数据库操作/SqlAlcheyObj/ProxyModel.py
@@ -45,7 +45,7 @@ class AvailableProxy(Base):
     ip = mapped_column(String(1024))
     counter = mapped_column(Integer)
     max_counter_ts = mapped_column(TIMESTAMP)
-    resp_code = mapped_column(TINYINT)
+    resp_code = mapped_column(Integer)
     available = mapped_column(TINYINT(1))
     latest_352_ts = mapped_column(TIMESTAMP)
     latest_used_ts = mapped_column(TIMESTAMP)
